오류수정

+++ src/main/java/com/sbs/cuni/controller/ArticleController.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.sbs.cuni.dto.Article;
 import com.sbs.cuni.dto.ArticleReply;
 import com.sbs.cuni.dto.Board;
+import com.sbs.cuni.dto.Member;
 import com.sbs.cuni.service.ArticleService;
 
 import lombok.extern.slf4j.Slf4j;
@@ -36,7 +37,9 @@ public class ArticleController {
 
 		// 게시물 가져올 때 댓글 개수도 가져오도록
 		param.put("extra__repliesCount", true);
-
+		// 게시물 작성자도 볼 수 있도록
+		param.put("extra__writer", true);
+		
 		if (param.containsKey("page") == false) {
 			param.put("page", "1");
 		}
@@ -52,7 +55,7 @@ public class ArticleController {
 	@RequestMapping("article/detail")
 	public String showDetail(@RequestParam(value = "id", defaultValue = "0") int id, Model model, long boardId) {
 		Board board = articleService.getBoard(boardId);
-
+		
 		model.addAttribute("board", board);
 
 		if (id == 0) {
@@ -80,12 +83,33 @@ public class ArticleController {
 	}
 
 	@RequestMapping("/article/add")
-	public String showAdd(long boardId, Model model) {
+	public String showAdd(long boardId, Model model, HttpServletRequest rq) {
 		Board board = articleService.getBoard(boardId);
-
+		String msg = "";
+		
+		String redirectUrl = "";
+		
+		Member member = (Member) rq.getAttribute("loginedMember");
+		
+		long permissionLevel = member.getPermissionLevel();
+		
+		if ( boardId == 1 ) {
+			if ( permissionLevel != 1 ) {
+				msg = "권한이 없습니다!";
+				
+				redirectUrl = "/";
+				
+				model.addAttribute("alertMsg", msg);
+				model.addAttribute("redirectUrl", redirectUrl);
+				
+				return "common/redirect";
+			}
+		}
+		
 		model.addAttribute("board", board);
+		
 
-		return "article/add";
+		return "/article/add";
 	}
 
 	@RequestMapping("/article/doAdd")
@@ -157,11 +181,27 @@ public class ArticleController {
 	@RequestMapping("/article/doDelete")
 	public String doDelete(Model model, @RequestParam Map<String, Object> param, HttpSession session, long id, long boardId) {
 		param.put("id", id);
+		
+		long loginedMemberId = (long)session.getAttribute("loginedMemberId");
 
+		Article article = articleService.getOne(param);
+		
+		String msg = "";
+		String resultCode = "";
+		
+		if(loginedMemberId != article.getMemberId()) {
+			msg = "권한이 없습니다!";
+			
+			model.addAttribute("alertMsg", msg);
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirect";
+		}
+		
 		Map<String, Object> deleteRs = articleService.delete(param);
 
-		String msg = (String) deleteRs.get("msg");
-		String resultCode = (String) deleteRs.get("resultCode");
+		msg = (String) deleteRs.get("msg");
+		resultCode = (String) deleteRs.get("resultCode");
 
 		if (resultCode.startsWith("S-")) {
 			String redirectUrl = "/article/list?boardId=" + boardId;
@@ -226,14 +266,31 @@ public class ArticleController {
 			e.printStackTrace();
 		}
 
+		Map<String, Object> rs = new HashMap<>();
+		String msg = "";
+		String resultCode = "";
+
+		long loginedMemberId = (long) session.getAttribute("loginedMemberId");
+
+		ArticleReply ar = articleService.getReply(param);
+
+		if ( loginedMemberId != ar.getMemberId() ) {
+			msg = "권한이 없습니다!";
+			resultCode = "F-5";
+
+			rs = Maps.of("msg", msg, "resultCode", resultCode);
+
+			return rs;
+		}
+		
 		param.put("id", id);
 
 		Map<String, Object> updateRs = articleService.updateReply(param);
 
-		String msg = (String) updateRs.get("msg");
-		String resultCode = (String) updateRs.get("resultCode");
+		msg = (String) updateRs.get("msg");
+		resultCode = (String) updateRs.get("resultCode");
 
-		Map<String, Object> rs = Maps.of("msg", msg, "resultCode", resultCode);
+		rs = Maps.of("msg", msg, "resultCode", resultCode);
 
 		return rs;
 	}
